+++ include/sw/dsp/sdr/sdr.hpp
@@ -9,6 +9,7 @@
 // Copyright (C) 2024-2026 Stillwater Supercomputing, Inc.
 // SPDX-License-Identifier: MIT
 
+#include <sw/dsp/sdr/agc.hpp>
 #include <sw/dsp/sdr/constellation.hpp>
 #include <sw/dsp/sdr/metrics.hpp>
 #include <sw/dsp/sdr/rrc.hpp>